requirements/dev.txt
beautifulsoup4==4.5.1
Django==1.9.9
django-extensions==1.7.3
djangorestframework==3.4.6
Pillow==3.3.1
pkg-resources==0.0.0
psycopg2==2.6.2
python-dateutil==2.5.3
redis==2.10.5
requests==2.11.1
selenium==2.53.6
six==1.10.0
